Checkpoint: adding accuracy to FFTW3 by using long double

+++ src/types.h
@@ -17,7 +17,7 @@ struct MatrixProfile {
 
 
 struct ComplexArray {
-    std::complex<double>* data = nullptr;
+    std::complex<long double>* data = nullptr;
     unsigned long length = 0;
 };
